Distance Sanity Check?
This will attempt to fix a weird bending (more of a detachment) from the upper body area from the lower body area which is caused by the NPC trying to shoot somewhere they physically can't(or maybe is unstreamed for them?). Anyways, I believe this may fix the issue a bit - however there are some additional bugs with the rotation of the aim that may cause this issue as well.